requirements.txt
paho-mqtt
tuya-iot-py-sdk
tinytuya